renamed CommonComponents to RefreshToggle
